different modifies
git-svn-id: https://n-exchange.googlecode.com/svn/trunk@115 37479ede-2111-a82d-6642-dbc92489a3ae

+++ src/main/java/it/sevenbits/dao/SearchVariantDao.java
@@ -21,7 +21,7 @@ public interface SearchVariantDao {
      * @param id
      * @return
      */
-    SearchVariant findById(Integer id);
+    SearchVariant findById(Long id);
 
     /**
      * возвращает  список всех подписчиков
